requirements.txt
# Core API framework
fastapi==0.115.6
uvicorn==0.34.0
sse-starlette==2.2.1
python-dotenv==1.0.1
pydantic==2.10.4
httpx==0.28.1

# Jenkins integration
python-jenkins==1.8.2

# LLM and agent framework
langchain==0.3.14
langchain-groq==0.2.4
langchain-community==0.3.14

# RAG - Vector store and embeddings
faiss-cpu==1.9.0.post1
sentence-transformers==3.3.1
numpy>=1.24.0

# Testing
pytest==8.3.4
pytest-asyncio==0.25.0
